move: avatar component to primitive
