+++ components/primitive/index.tsx
@@ -6,6 +6,19 @@ import { Overlay } from 'components/primitive/Overlay';
 import { Panel } from 'components/primitive/Panel';
 import { Status } from 'components/primitive/Status';
 import { List, ListItem } from 'components/primitive/List';
+import { Avatar } from 'components/primitive/Avatar';
 import { Link } from 'components/primitive/Link';
 
-export { Box, Flex, Text, Grid, Overlay, Panel, Status, List, ListItem, Link };
+export {
+  Box,
+  Flex,
+  Text,
+  Grid,
+  Overlay,
+  Panel,
+  Status,
+  List,
+  ListItem,
+  Link,
+  Avatar,
+};
